+++ tests/mobc.rs
@@ -187,7 +187,7 @@ fn test_drop_on_broken() {
     rt.block_on(async {
         let pool = Pool::new(handler);
 
-        drop(pool.get().await.ok().unwrap());
+        assert!(pool.get().await.is_err());
         delay_for(Duration::from_secs(1)).await;
         assert!(DROPPED.load(Ordering::SeqCst));
         Ok::<(), Error<TestError>>(())
@@ -228,7 +228,7 @@ fn test_invalid_conn() {
     rt.block_on(async {
         let pool = Pool::builder().max_open(1).build(Handler);
 
-        drop(pool.get().await.ok().unwrap());
+        assert!(pool.get().await.is_err());
         delay_for(Duration::from_secs(1)).await;
         assert!(DROPPED.load(Ordering::SeqCst));
         assert_eq!(1_u64, pool.state().await.connections);